+++ tests/callbacks/test_callback_hook_outputs.py
@@ -65,48 +65,6 @@ def test_train_step_no_return(tmpdir, single_cb: bool):
     trainer.fit(model)
 
 
-def test_on_val_epoch_end_outputs(tmpdir):
-
-    class CB(Callback):
-
-        def on_validation_epoch_end(self, trainer, pl_module, outputs):
-            if trainer.running_sanity_check:
-                assert len(outputs) == trainer.num_sanity_val_batches[0]
-            else:
-                assert len(outputs) == trainer.num_val_batches[0]
-
-    model = BoringModel()
-
-    trainer = Trainer(
-        callbacks=CB(),
-        default_root_dir=tmpdir,
-        limit_train_batches=2,
-        limit_val_batches=2,
-        max_epochs=1,
-        weights_summary=None,
-    )
-
-    trainer.fit(model)
-
-
-def test_on_test_epoch_end_outputs(tmpdir):
-
-    class CB(Callback):
-
-        def on_test_epoch_end(self, trainer, pl_module, outputs):
-            assert len(outputs) == trainer.num_test_batches[0]
-
-    model = BoringModel()
-
-    trainer = Trainer(
-        callbacks=CB(),
-        default_root_dir=tmpdir,
-        weights_summary=None,
-    )
-
-    trainer.test(model)
-
-
 def test_free_memory_on_eval_outputs(tmpdir):
 
     class CB(Callback):