+++ tests/trainer/logging_/test_logger_connector.py
@@ -681,10 +681,10 @@ def test_metrics_reset(tmpdir):
         def on_train_epoch_end(self):
             self._assert_epoch_end('train')
 
-        def on_validation_epoch_end(self, outputs):
+        def on_validation_epoch_end(self):
             self._assert_epoch_end('val')
 
-        def on_test_epoch_end(self, outputs):
+        def on_test_epoch_end(self):
             self._assert_epoch_end('test')
 
     def _assert_called(model, stage):